controlstyle_ui/requirements.txt
accelerate
datasets==2.10.1
diffusers
gradio==3.20.0
huggingface_hub
numpy
packaging==23.0
Pillow==9.4.0
torch
torchvision
tqdm==4.64.1
transformers
omegaconf==2.2.3
